requirements.txt
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.24.0
torch
safetensors>=0.4.0
pillow>=10.0.0
numpy>=1.24.0
